Close #578: Rename `extras.core.syntax.string` to `extras.core.syntax.strings`
