+++ modules/extras-core/shared/src/main/scala-2/extras/core/syntax/all.scala
@@ -1,0 +1,7 @@
+package extras.core.syntax
+
+/** @author Kevin Lee
+  * @since 2022-06-25
+  */
+trait all extends strings
+object all extends all
