--- modules/extras-core/shared/src/main/scala/extras/core/syntax/package.scala
@@ -1,10 +1,0 @@
-package extras.core
-
-/** @author Kevin Lee
-  * @since 2022-06-25
-  */
-package object syntax {
-  object string extends StringSyntax
-
-  object all extends StringSyntax
-}
