+++ modules/extras-core/shared/src/main/scala/extras/core/syntax/string.scala
@@ -1,0 +1,7 @@
+package extras.core.syntax
+
+/** @author Kevin Lee
+  * @since 2022-06-25
+  */
+@deprecated(message = "Use strings instead", since = "0.50.0")
+object string extends strings
